use changes in SOURCE as the trigger to display activation msgs
Before I was just checking if the VERSION changed.  This works but
hides a little info.  If the version you have activated is the same
as a new project you go to, you won't be alerted that a new version
file is in effect.

By detecting source changes, you are alerted each time a new SOURCE
takes effect.