rework the activation script and CLI
This changes the activation code to remove the running bash commands
from the rb CLI.  Now rb always sets the env based on params.  This
change meant having rb add two function to the shell, one for the
activation script and one for the CLI.

This also reorganizes the scripts and move CLI cmd scripts into
their own file for ease in maintenance.  The CLI cmds are just
helpers and don't affect the env.

* Closes #12
* Closes #5