resolve ruby version symlinks when applying to the env
This ensures that an actual version path (not a symlinked path)
is always used in the ENV vars.  This means that symlinking ruby
versions will have no effect on the path related ENV vars.  The
symlinks will only be used for referencing versions.

This is good b/c it normalizes things.  Plus you can safely treat
symlinks as references only.

Finally, this adds an additional ENV var: `RUBY_VERSION`.  This is
necessary to properly remove path entries as `RB_RUBY_VERSION` may
not match the path b/c it may be unresolved.  I have wanted to set
this var anyway and this just forced the issue.