Use `dirname` for getting rb's install directory
On Fedora systems, a command like the following:

```bash
ls -la $(readlink `which rb`)/..
```

fails. It says `ls: cannot access <path>: Not a directory`. I
believe it is complaining about the fact that the `rb` bin is
not a directory. I can't find exactly what this is, but it seems
to be some strict path handling.

Anyway, this changes the couple of places I've found in `rb` that
use the pattern of calling readling and appending `..`. I switched
to using `dirname` instead which is works with Fedora and Mac OS
X.