switch to finding default `.ruby-version` file in the rubies dir
This switches to looking for the default ruby version settings file
in the rubies root dir instead of the rb root dir.

The problem with storing the default version in the rb root dir is
that when you upgrade rb, that dir is removed and recreated.  Therefore
your ruby version file is destroyed with it and you have to manually
recreate it.

This moves the file to live with the rubies in the rubies root dir.
This dir will not be destroyed when you upgrade rb so your default
ruby version will remain intact between upgrades.