make the install and release scripts more robust
This updates the install script to no longer wipe the `.rb` dir
but instead keep the history of previously installed versions.
This allows any custom files added to the `.rb` dir to persist
across new installs.

This updates the release script to actually halt on fault
conditions and makes the script more fault tolerant.

These came out of working on Chpg and needing the features there.